% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonGroup.R
\name{buttonGroup}
\alias{buttonGroup}
\title{Creates an example input binding.
This binding consists of a simple div that will generate a random number each time it is clicked}
\usage{
buttonGroup(inputId, choices, selected = NA)
}
\arguments{
\item{inputId}{id for the component}

\item{choices}{list with choices that should be available for the buttongroup}

\item{selected}{list with values that should already be selected}
}
\description{
Creates an example input binding.
This binding consists of a simple div that will generate a random number each time it is clicked
}
